backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
python-dotenv>=1.0.0
pymongo>=4.0.0
sentence-transformers==2.7.0
transformers==4.40.2
huggingface-hub==0.23.4
anthropic>=0.18.0
pymupdf>=1.23.0
python-docx>=1.0.0
openpyxl>=3.1.0
python-pptx>=0.6.0
numpy>=1.24.0
certifi>=2023.0.0
